requirement.txt
Flask==2.3.3
flask-socketio==5.3.6
crewai==0.28.0
python-dotenv==1.0.0
pandas==2.0.3
python-docx==0.8.11
Werkzeug==2.3.7
openpyxl==3.1.2
langchain==0.1.0
langchain-openai==0.0.2